+++ tencentcloud/dlc/v20210125/client.go
@@ -324,6 +324,77 @@ func (c *Client) AlterDMSTableWithContext(ctx context.Context, request *AlterDMS
     return
 }
 
+func NewAssignMangedTablePropertiesRequest() (request *AssignMangedTablePropertiesRequest) {
+    request = &AssignMangedTablePropertiesRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    
+    request.Init().WithApiInfo("dlc", APIVersion, "AssignMangedTableProperties")
+    
+    
+    return
+}
+
+func NewAssignMangedTablePropertiesResponse() (response *AssignMangedTablePropertiesResponse) {
+    response = &AssignMangedTablePropertiesResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    } 
+    return
+
+}
+
+// AssignMangedTableProperties
+// 分配原生表表属性
+//
+// 可能返回的错误码:
+//  FAILEDOPERATION = "FailedOperation"
+//  INTERNALERROR = "InternalError"
+//  INTERNALERROR_INTERNALSYSTEMEXCEPTION = "InternalError.InternalSystemException"
+//  INVALIDPARAMETER = "InvalidParameter"
+//  INVALIDPARAMETER_INVALIDCOLUMNNAMELENGTH = "InvalidParameter.InvalidColumnNameLength"
+//  INVALIDPARAMETER_INVALIDCOLUMNNUMBER = "InvalidParameter.InvalidColumnNumber"
+//  INVALIDPARAMETER_INVALIDDECIMALTYPE = "InvalidParameter.InvalidDecimalType"
+//  INVALIDPARAMETER_INVALIDTABLENAMELENGTH = "InvalidParameter.InvalidTableNameLength"
+//  INVALIDPARAMETER_PARAMETERNOTFOUNDORBENONE = "InvalidParameter.ParameterNotFoundOrBeNone"
+//  RESOURCEUNAVAILABLE = "ResourceUnavailable"
+//  UNAUTHORIZEDOPERATION = "UnauthorizedOperation"
+//  UNSUPPORTEDOPERATION = "UnsupportedOperation"
+func (c *Client) AssignMangedTableProperties(request *AssignMangedTablePropertiesRequest) (response *AssignMangedTablePropertiesResponse, err error) {
+    return c.AssignMangedTablePropertiesWithContext(context.Background(), request)
+}
+
+// AssignMangedTableProperties
+// 分配原生表表属性
+//
+// 可能返回的错误码:
+//  FAILEDOPERATION = "FailedOperation"
+//  INTERNALERROR = "InternalError"
+//  INTERNALERROR_INTERNALSYSTEMEXCEPTION = "InternalError.InternalSystemException"
+//  INVALIDPARAMETER = "InvalidParameter"
+//  INVALIDPARAMETER_INVALIDCOLUMNNAMELENGTH = "InvalidParameter.InvalidColumnNameLength"
+//  INVALIDPARAMETER_INVALIDCOLUMNNUMBER = "InvalidParameter.InvalidColumnNumber"
+//  INVALIDPARAMETER_INVALIDDECIMALTYPE = "InvalidParameter.InvalidDecimalType"
+//  INVALIDPARAMETER_INVALIDTABLENAMELENGTH = "InvalidParameter.InvalidTableNameLength"
+//  INVALIDPARAMETER_PARAMETERNOTFOUNDORBENONE = "InvalidParameter.ParameterNotFoundOrBeNone"
+//  RESOURCEUNAVAILABLE = "ResourceUnavailable"
+//  UNAUTHORIZEDOPERATION = "UnauthorizedOperation"
+//  UNSUPPORTEDOPERATION = "UnsupportedOperation"
+func (c *Client) AssignMangedTablePropertiesWithContext(ctx context.Context, request *AssignMangedTablePropertiesRequest) (response *AssignMangedTablePropertiesResponse, err error) {
+    if request == nil {
+        request = NewAssignMangedTablePropertiesRequest()
+    }
+    
+    if c.GetCredential() == nil {
+        return nil, errors.New("AssignMangedTableProperties require credential")
+    }
+
+    request.SetContext(ctx)
+    
+    response = NewAssignMangedTablePropertiesResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewAttachUserPolicyRequest() (request *AttachUserPolicyRequest) {
     request = &AttachUserPolicyRequest{
         BaseRequest: &tchttp.BaseRequest{},