+++ tencentcloud/ess/v20201111/models.go
@@ -1582,6 +1582,9 @@ type CreateDocumentRequestParams struct {
 	FileNames []*string `json:"FileNames,omitnil" name:"FileNames"`
 
 	// 电子文档的填写控件的填充内容。具体方式可以参考[FormField](https://qian.tencent.com/developers/companyApis/dataTypes/#formfield)结构体的定义。
+	// <ul>
+	// <li>支持自动签传递印章，可通过指定自动签控件id，指定印章id来完成</li>
+	// </ul>
 	FormFields []*FormField `json:"FormFields,omitnil" name:"FormFields"`
 
 	// 是否为预览模式，取值如下：
@@ -1623,6 +1626,9 @@ type CreateDocumentRequest struct {
 	FileNames []*string `json:"FileNames,omitnil" name:"FileNames"`
 
 	// 电子文档的填写控件的填充内容。具体方式可以参考[FormField](https://qian.tencent.com/developers/companyApis/dataTypes/#formfield)结构体的定义。
+	// <ul>
+	// <li>支持自动签传递印章，可通过指定自动签控件id，指定印章id来完成</li>
+	// </ul>
 	FormFields []*FormField `json:"FormFields,omitnil" name:"FormFields"`
 
 	// 是否为预览模式，取值如下：