+++ tencentcloud/lcic/v20220817/models.go
@@ -769,8 +769,8 @@ type CreateDocumentRequestParams struct {
 	Owner *string `json:"Owner,omitnil" name:"Owner"`
 
 	// 转码类型，可以有如下取值：
-	// 0 无需转码（默认）
-	// 1 需要转码的文档，ppt，pptx，pdf，doc，docx
+	// 0 无需转码（默认），bmp，jpg，jpeg，png，gif
+	// 1 需要转码的文档，ppt，pptx，pdf，doc，docx，xls，xlsx
 	// 2 需要转码的视频，mp4，3pg，mpeg，avi，flv，wmv，rm，h264等
 	// 2 需要转码的音频，mp3，wav，wma，aac，flac，opus
 	TranscodeType *uint64 `json:"TranscodeType,omitnil" name:"TranscodeType"`
@@ -810,8 +810,8 @@ type CreateDocumentRequest struct {
 	Owner *string `json:"Owner,omitnil" name:"Owner"`
 
 	// 转码类型，可以有如下取值：
-	// 0 无需转码（默认）
-	// 1 需要转码的文档，ppt，pptx，pdf，doc，docx
+	// 0 无需转码（默认），bmp，jpg，jpeg，png，gif
+	// 1 需要转码的文档，ppt，pptx，pdf，doc，docx，xls，xlsx
 	// 2 需要转码的视频，mp4，3pg，mpeg，avi，flv，wmv，rm，h264等
 	// 2 需要转码的音频，mp3，wav，wma，aac，flac，opus
 	TranscodeType *uint64 `json:"TranscodeType,omitnil" name:"TranscodeType"`