+++ tencentcloud/ssl/v20191205/models.go
@@ -621,6 +621,22 @@ type Certificates struct {
 	// 是否允许下载
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	AllowDownload *bool `json:"AllowDownload,omitnil" name:"AllowDownload"`
+
+	// 证书域名是否全部在DNSPOD托管解析
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IsDNSPODResolve *bool `json:"IsDNSPODResolve,omitnil" name:"IsDNSPODResolve"`
+
+	// 是否是权益点购买的证书
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IsPackage *bool `json:"IsPackage,omitnil" name:"IsPackage"`
+
+	// 是否存在私钥密码
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	KeyPasswordCustomFlag *bool `json:"KeyPasswordCustomFlag,omitnil" name:"KeyPasswordCustomFlag"`
+
+	// 支持下载的WEB服务器类型： nginx、apache、iis、tomcat、jks、root、other
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	SupportDownloadType *SupportDownloadType `json:"SupportDownloadType,omitnil" name:"SupportDownloadType"`
 }
 
 // Predefined struct for user
@@ -5749,6 +5765,29 @@ type SubmittedData struct {
 	VerifyType *string `json:"VerifyType,omitnil" name:"VerifyType"`
 }
 
+type SupportDownloadType struct {
+	// 是否可以下载nginx可用格式
+	NGINX *bool `json:"NGINX,omitnil" name:"NGINX"`
+
+	// 是否可以下载apache可用格式
+	APACHE *bool `json:"APACHE,omitnil" name:"APACHE"`
+
+	// 是否可以下载tomcat可用格式
+	TOMCAT *bool `json:"TOMCAT,omitnil" name:"TOMCAT"`
+
+	// 是否可以下载iis可用格式
+	IIS *bool `json:"IIS,omitnil" name:"IIS"`
+
+	// 是否可以下载JKS可用格式
+	JKS *bool `json:"JKS,omitnil" name:"JKS"`
+
+	// 是否可以下载其他格式
+	OTHER *bool `json:"OTHER,omitnil" name:"OTHER"`
+
+	// 是否可以下载根证书
+	ROOT *bool `json:"ROOT,omitnil" name:"ROOT"`
+}
+
 type SyncTaskBindResourceResult struct {
 	// 任务ID
 	TaskId *string `json:"TaskId,omitnil" name:"TaskId"`
@@ -5999,13 +6038,13 @@ type TkeSecretDetail struct {
 
 // Predefined struct for user
 type UpdateCertificateInstanceRequestParams struct {
-	// 一键更新原证书ID
+	// 一键更新原证书ID， 查询绑定该证书的云资源然后进行证书更新
 	OldCertificateId *string `json:"OldCertificateId,omitnil" name:"OldCertificateId"`
 
-	// 需要部署的资源类型，参数值可选：clb、cdn、waf、live、ddos、teo、apigateway、vod、tke、tcb
+	// 需要部署的资源类型，参数值可选（小写）：clb、cdn、waf、live、ddos、teo、apigateway、vod、tke、tcb、tse
 	ResourceTypes []*string `json:"ResourceTypes,omitnil" name:"ResourceTypes"`
 
-	// 一键更新新证书ID
+	// 一键更新新证书ID，不传该则证书公钥和私钥必传
 	CertificateId *string `json:"CertificateId,omitnil" name:"CertificateId"`
 
 	// 需要部署的地域列表（废弃）
@@ -6013,68 +6052,68 @@ type UpdateCertificateInstanceRequestParams struct {
 	// Deprecated: Regions is deprecated.
 	Regions []*string `json:"Regions,omitnil" name:"Regions"`
 
-	// 云资源需要部署的地域列表
+	// 云资源需要部署的地域列表，支持地域的云资源类型必传，如：clb、tke、apigateway、waf、tcb、tse等
 	ResourceTypesRegions []*ResourceTypeRegions `json:"ResourceTypesRegions,omitnil" name:"ResourceTypesRegions"`
 
 	// 证书公钥， 若上传证书公钥， 则CertificateId不用传
 	CertificatePublicKey *string `json:"CertificatePublicKey,omitnil" name:"CertificatePublicKey"`
 
-	// 证书私钥，若上传证书公钥， 则证书私钥必填
+	// 证书私钥，若上传证书公钥， 则CertificateId不用传
 	CertificatePrivateKey *string `json:"CertificatePrivateKey,omitnil" name:"CertificatePrivateKey"`
 
-	// 旧证书是否忽略到期提醒  0:不忽略通知。1:忽略通知
+	// 旧证书是否忽略到期提醒  0:不忽略通知。1:忽略通知，忽略OldCertificateId到期提醒
 	ExpiringNotificationSwitch *uint64 `json:"ExpiringNotificationSwitch,omitnil" name:"ExpiringNotificationSwitch"`
 
-	// 相同的证书是否允许重复上传，若上传证书公钥， 则可以配置该参数
+	// 相同的证书是否允许重复上传，若选择上传证书， 则可以配置该参数
 	Repeatable *bool `json:"Repeatable,omitnil" name:"Repeatable"`
 
-	// 是否允许下载，若上传证书公钥， 则可以配置该参数
+	// 是否允许下载，若选择上传证书， 则可以配置该参数
 	AllowDownload *bool `json:"AllowDownload,omitnil" name:"AllowDownload"`
 
-	// 标签列表，若上传证书公钥， 则可以配置该参数
+	// 标签列表，若选择上传证书， 则可以配置该参数
 	Tags []*Tags `json:"Tags,omitnil" name:"Tags"`
 
-	// 项目 ID，若上传证书公钥， 则可以配置该参数
+	// 项目 ID，若选择上传证书， 则可以配置该参数
 	ProjectId *uint64 `json:"ProjectId,omitnil" name:"ProjectId"`
 }
 
 type UpdateCertificateInstanceRequest struct {
 	*tchttp.BaseRequest
 	
-	// 一键更新原证书ID
+	// 一键更新原证书ID， 查询绑定该证书的云资源然后进行证书更新
 	OldCertificateId *string `json:"OldCertificateId,omitnil" name:"OldCertificateId"`
 
-	// 需要部署的资源类型，参数值可选：clb、cdn、waf、live、ddos、teo、apigateway、vod、tke、tcb
+	// 需要部署的资源类型，参数值可选（小写）：clb、cdn、waf、live、ddos、teo、apigateway、vod、tke、tcb、tse
 	ResourceTypes []*string `json:"ResourceTypes,omitnil" name:"ResourceTypes"`
 
-	// 一键更新新证书ID
+	// 一键更新新证书ID，不传该则证书公钥和私钥必传
 	CertificateId *string `json:"CertificateId,omitnil" name:"CertificateId"`
 
 	// 需要部署的地域列表（废弃）
 	Regions []*string `json:"Regions,omitnil" name:"Regions"`
 
-	// 云资源需要部署的地域列表
+	// 云资源需要部署的地域列表，支持地域的云资源类型必传，如：clb、tke、apigateway、waf、tcb、tse等
 	ResourceTypesRegions []*ResourceTypeRegions `json:"ResourceTypesRegions,omitnil" name:"ResourceTypesRegions"`
 
 	// 证书公钥， 若上传证书公钥， 则CertificateId不用传
 	CertificatePublicKey *string `json:"CertificatePublicKey,omitnil" name:"CertificatePublicKey"`
 
-	// 证书私钥，若上传证书公钥， 则证书私钥必填
+	// 证书私钥，若上传证书公钥， 则CertificateId不用传
 	CertificatePrivateKey *string `json:"CertificatePrivateKey,omitnil" name:"CertificatePrivateKey"`
 
-	// 旧证书是否忽略到期提醒  0:不忽略通知。1:忽略通知
+	// 旧证书是否忽略到期提醒  0:不忽略通知。1:忽略通知，忽略OldCertificateId到期提醒
 	ExpiringNotificationSwitch *uint64 `json:"ExpiringNotificationSwitch,omitnil" name:"ExpiringNotificationSwitch"`
 
-	// 相同的证书是否允许重复上传，若上传证书公钥， 则可以配置该参数
+	// 相同的证书是否允许重复上传，若选择上传证书， 则可以配置该参数
 	Repeatable *bool `json:"Repeatable,omitnil" name:"Repeatable"`
 
-	// 是否允许下载，若上传证书公钥， 则可以配置该参数
+	// 是否允许下载，若选择上传证书， 则可以配置该参数
 	AllowDownload *bool `json:"AllowDownload,omitnil" name:"AllowDownload"`
 
-	// 标签列表，若上传证书公钥， 则可以配置该参数
+	// 标签列表，若选择上传证书， 则可以配置该参数
 	Tags []*Tags `json:"Tags,omitnil" name:"Tags"`
 
-	// 项目 ID，若上传证书公钥， 则可以配置该参数
+	// 项目 ID，若选择上传证书， 则可以配置该参数
 	ProjectId *uint64 `json:"ProjectId,omitnil" name:"ProjectId"`
 }
 
@@ -6116,6 +6155,10 @@ type UpdateCertificateInstanceResponseParams struct {
 
 	// 部署状态，1表示部署成功，0表示部署失败
 	DeployStatus *int64 `json:"DeployStatus,omitnil" name:"DeployStatus"`
+
+	// 更新异步创建任务进度详情
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	UpdateSyncProgress []*UpdateSyncProgress `json:"UpdateSyncProgress,omitnil" name:"UpdateSyncProgress"`
 
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 	RequestId *string `json:"RequestId,omitnil" name:"RequestId"`
@@ -6374,6 +6417,38 @@ type UpdateRecordInfo struct {
 
 	// 最后一次更新时间
 	UpdateTime *string `json:"UpdateTime,omitnil" name:"UpdateTime"`
+}
+
+type UpdateSyncProgress struct {
+	// 资源类型
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ResourceType *string `json:"ResourceType,omitnil" name:"ResourceType"`
+
+	// 地域结果列表
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	UpdateSyncProgressRegions []*UpdateSyncProgressRegion `json:"UpdateSyncProgressRegions,omitnil" name:"UpdateSyncProgressRegions"`
+
+	// 异步更新进度状态：0， 待处理， 1 已处理， 3 处理中
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Status *int64 `json:"Status,omitnil" name:"Status"`
+}
+
+type UpdateSyncProgressRegion struct {
+	// 资源类型
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Region *string `json:"Region,omitnil" name:"Region"`
+
+	// 总数
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	TotalCount *int64 `json:"TotalCount,omitnil" name:"TotalCount"`
+
+	// 执行完成数量
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OffsetCount *int64 `json:"OffsetCount,omitnil" name:"OffsetCount"`
+
+	// 异步更新进度状态：0， 待处理， 1 已处理， 3 处理中
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Status *int64 `json:"Status,omitnil" name:"Status"`
 }
 
 // Predefined struct for user