+++ tencentcloud/vod/v20180717/models.go
@@ -23007,12 +23007,37 @@ func (r *SetDrmKeyProviderInfoResponse) FromJsonString(s string) error {
 
 // Predefined struct for user
 type SetVodDomainCertificateRequestParams struct {
+	// 域名。
+	Domain *string `json:"Domain,omitnil" name:"Domain"`
 
+	// 本次接口调用的操作类型，可选值：
+	// <li>Set：设置域名的 HTTPS 所需证书；</li>
+	// <li>Clear: 清空域名 HTTPS 配置，清空后该域名无法支持 HTTPS 访问。</li>
+	Operation *string `json:"Operation,omitnil" name:"Operation"`
+
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
+	SubAppId *uint64 `json:"SubAppId,omitnil" name:"SubAppId"`
+
+	// [腾讯云SSL](/document/product/400/7572) 上传的证书 ID。清空域名 HTTPS 配置时无需填写该字段。
+	CertID *string `json:"CertID,omitnil" name:"CertID"`
 }
 
 type SetVodDomainCertificateRequest struct {
 	*tchttp.BaseRequest
 	
+	// 域名。
+	Domain *string `json:"Domain,omitnil" name:"Domain"`
+
+	// 本次接口调用的操作类型，可选值：
+	// <li>Set：设置域名的 HTTPS 所需证书；</li>
+	// <li>Clear: 清空域名 HTTPS 配置，清空后该域名无法支持 HTTPS 访问。</li>
+	Operation *string `json:"Operation,omitnil" name:"Operation"`
+
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
+	SubAppId *uint64 `json:"SubAppId,omitnil" name:"SubAppId"`
+
+	// [腾讯云SSL](/document/product/400/7572) 上传的证书 ID。清空域名 HTTPS 配置时无需填写该字段。
+	CertID *string `json:"CertID,omitnil" name:"CertID"`
 }
 
 func (r *SetVodDomainCertificateRequest) ToJsonString() string {
@@ -23027,7 +23052,10 @@ func (r *SetVodDomainCertificateRequest) FromJsonString(s string) error {
 	if err := json.Unmarshal([]byte(s), &f); err != nil {
 		return err
 	}
-	
+	delete(f, "Domain")
+	delete(f, "Operation")
+	delete(f, "SubAppId")
+	delete(f, "CertID")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "SetVodDomainCertificateRequest has unknown keys!", "")
 	}